hw4/webapp/requirements.txt
Flask>=1.0.0
cloudant==2.13.0
wiotp-sdk
joblib
scikit-learn